发布问题

+++ src/main/java/com/sf/skytalk/controller/QuestionController.java
@@ -15,7 +15,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable Integer id
+    public String question(@PathVariable Long id
             , Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         model.addAttribute("question",questionDTO);
